+++ TechnicalRadiation.Services/NewsService.cs
@@ -64,6 +64,8 @@ namespace TechnicalRadiation.Services
 
         public void DeleteNewsById(int id)
         {
+            if (id < 1) { throw new ArgumentOutOfRangeException("Id should not be lower than 1"); }
+            if (FakeDatabase.NewsItems.Count < id) { throw new ResourceNotFoundException($"News item with id {id} was not found"); }
             _newsRepo.DeleteNewsById(id);
         }
     }
